SequenceDS inherits from Sequence, not StructureDS SequenceDS, StructureDS implement StructureEnhanced.
